Add files via upload
Contains code for the webscraping (uncleaned). The final dataset called 'hackathons_webscraped.csv' is available and contains the columns - Subtitles, sponsors, Tags, Judges, Judging Criteria, Eligibility, Requirements, and URL, for all 7048 rows (out of 7053 original, some had errors pulling the data but since this is a very small number, it should not affect the analysis).